+++ src/worker/core/season/index.ts
@@ -1,3 +1,4 @@
+import addClinchedPlayoffs from "./addClinchedPlayoffs";
 import doAwards from "./doAwards";
 import genPlayoffSeries from "./genPlayoffSeries";
 import getAwardCandidates from "./getAwardCandidates";
@@ -9,6 +10,7 @@ import setSchedule from "./setSchedule";
 import updateOwnerMood from "./updateOwnerMood";
 
 export default {
+	addClinchedPlayoffs,
 	doAwards,
 	genPlayoffSeries,
 	getAwardCandidates,